Consolidate baseline helpers
