+++ baseline.py
@@ -2,10 +2,12 @@ import numpy as np
 import logging
 import pandas as pd
 
+from radon import baseline as _baseline
 import baseline_utils
-from baseline_utils import subtract_baseline_dataframe
 
-__all__ = ["rate_histogram", "subtract_baseline", "subtract_baseline_dataframe"]
+subtract_baseline = baseline_utils.apply_baseline_correction
+
+__all__ = ["rate_histogram", "subtract_baseline", "apply_baseline_correction"]
 
 
 def rate_histogram(df, bins):
@@ -17,7 +19,7 @@ def rate_histogram(df, bins):
     """
     if df.empty:
         return np.zeros(len(bins) - 1, dtype=float), 0.0
-    ts = baseline_utils._to_datetime64(df["timestamp"])
+    ts = _baseline._to_datetime64(df["timestamp"])
     ts_int = ts.view("int64")
     live = float((ts_int[-1] - ts_int[0]) / 1e9)
     hist_src = df.get("subtracted_adc_hist", df["adc"]).to_numpy()
@@ -25,18 +27,3 @@ def rate_histogram(df, bins):
     if live <= 0:
         return np.zeros_like(hist, dtype=float), live
     return hist / live, live
-
-
-def subtract_baseline(df_analysis, df_full, bins, t_base0, t_base1,
-                         mode="all", live_time_analysis=None):
-    """Wrapper for :func:`baseline_utils.subtract_baseline_dataframe`."""
-
-    return subtract_baseline_dataframe(
-        df_analysis,
-        df_full,
-        bins,
-        t_base0,
-        t_base1,
-        mode=mode,
-        live_time_analysis=live_time_analysis,
-    )
